requirements.txt
streamlit==1.33.0
tensorflow==2.19.0
numpy==1.26.4
pillow==10.2.0
opencv-python-headless==4.8.1.78
pandas==2.1.4
matplotlib==3.8.4